Added Delete confirmation for todo

+++ frontend/src/components/TodoItem.jsx
@@ -34,7 +34,9 @@ function TodoItem({ todo }) {
     };
 
     const onDeleteTodo = (id) => {
-        dispatch(deleteTodo(id));
+        if (window.confirm('Are you sure?') === true) {
+            dispatch(deleteTodo(id));
+        }
     };
 
     const toggleFavourite = () => {
